initial approach
As the parent is going to add the numbers received from the
children, we can encode the numbers as the number of signals sent
to the parent, and the total number of signasl will be the final
total.  The parent doesn't know the individual numbers sent by
the children... but it can count the total number of signals and
the count will equal the sum.

Signed-off-by: Luis Colorado <luiscoloradourcola@gmail.com>